Clean up empty bindings.
When adding/removing lots of bindings, one ends up with many leftover properties on this._events; this cleans them up by deleting the property when unbinding (if it is empty). With this in place, a bug is created whereby a callback may unbind itself while trigger is looping through. If that happens, the for statement's .length check will fail since the index is no longer there, so we calculate the length once beforehand.